+++ src/main/java/eu/iv4xr/framework/extensions/occ/EmotionMemory.java
@@ -6,7 +6,7 @@ public class EmotionMemory {
 
     List<Emotion> memory = new LinkedList<>();
 
-    public int memoryHorizon = 1000;    
+    public long memoryHorizon = 1000;    
 
     /**
      * Add this emotion e to the memory.
@@ -19,8 +19,8 @@ public class EmotionMemory {
     /**
      * Remove memories which are too old.
      */
-    public void cleanup(int currentTime) {
-        int oldestAllowed = Math.max(0, currentTime - memoryHorizon);
+    public void cleanup(long currentTime) {
+        long oldestAllowed = Math.max(0, currentTime - memoryHorizon);
         memory.removeIf(e -> e.t0 < oldestAllowed);
     }
